feat: Board Delete

+++ src/boards/boards.service.ts
@@ -40,8 +40,10 @@ export class BoardsService {
     return this.boardRepository.createBoard(createBoardDto, user);
   }
 
-  async deleteBoard(id: number): Promise<void> {
-    const result = await this.boardRepository.delete({ id });
+  async deleteBoard(id: number, user: User): Promise<void> {
+    console.log(user);
+
+    const result = await this.boardRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Can't find Board with id ${id}`);
     }
